+++ analyzers/__init__.py
@@ -1,0 +1,7 @@
+from error_stat import AnalyzeError
+from coverage_modes import AnalyzeCoverageModes
+from coverage_stop_area import AnalyzeCoverageStopArea
+from requests_calls import AnalyzeRequest
+from token_stat import AnalyzeToken
+from users_sql import AnalyseUsersSql
+from analyzer import  Analyzer
