+++ analyzers/coverage_stop_areas.py
@@ -1,5 +1,4 @@
-import json
-from analyzer import Analyzer
+from analyzers.analyzer import Analyzer
 from datetime import datetime
 
 
@@ -72,10 +71,7 @@ class AnalyzeCoverageStopAreas(Analyzer):
 
     def get_data(self):
         files = self.get_files_to_analyze()
-        rdd = self.spark_session.sparkContext.textFile(','.join(files)).map(
-            # json to dict
-            lambda stat: json.loads(stat)
-        )
+        rdd = self.load_data(files, rdd_mode=True)
         return self.collect_data_from_df(rdd)
 
     def truncate_and_insert(self, data):
