+++ analyzers/error_stats.py
@@ -1,5 +1,5 @@
 from pyspark.sql.functions import when, from_unixtime
-from analyzer import Analyzer
+from analyzers.analyzer import Analyzer
 
 
 class AnalyzeErrors(Analyzer):
@@ -20,7 +20,7 @@ class AnalyzeErrors(Analyzer):
 
     def get_data(self):
         files = self.get_files_to_analyze()
-        df = self.spark_session.read.json(files)
+        df = self.load_data(files)
         return self.collect_data_from_df(df)
 
     def truncate_and_insert(self, data):
