+++ analyzers/requests_calls.py
@@ -1,4 +1,4 @@
-from analyzer import Analyzer
+from analyzers.analyzer import Analyzer
 from pyspark.sql.functions import when, from_unixtime, lit
 
 
@@ -43,7 +43,7 @@ class AnalyzeRequest(Analyzer):
 
     def get_data(self):
         files = self.get_files_to_analyze()
-        df = self.spark_session.read.json(files)
+        df = self.load_data(files)
         return self.collect_data_from_df(df)
 
     def truncate_and_insert(self, data):
