+++ includes/utils.py
@@ -1,16 +1,18 @@
 import os
 from datetime import datetime
-from analyzers import token_stat, users_sql, error_stats, requests_calls,  coverage_stop_areas
+from analyzers import AnalyseUsersSql, AnalyzeToken, AnalyzeRequest,\
+    AnalyzeError, AnalyzeCoverageStopArea, AnalyzeCoverageModes
 from itertools import chain, islice
 
 
 def analyzer_value(value):
     analyzers = {
-        "token_stats": token_stat.AnalyzeToken,
-        "users": users_sql.AnalyseUsersSql,
-        "requests_calls": requests_calls.AnalyzeRequest,
-        "error_stats": error_stats.AnalyzeErrors,
-        "coverage_stop_areas": coverage_stop_areas.AnalyzeCoverageStopAreas
+        "token_stats": AnalyzeToken,
+        "users": AnalyseUsersSql,
+        "requests_calls": AnalyzeRequest,
+        "error_stat": AnalyzeError,
+        "coverage_stop_area": AnalyzeCoverageStopArea,
+        "coverage_modes": AnalyzeCoverageModes
     }
     lower_value = value.lower()
     if lower_value not in analyzers:
