+++ tests/coverage_stop_areas_test.py
@@ -1,6 +1,6 @@
 import pytest
 from datetime import date, datetime
-from analyzers.coverage_stop_areas import AnalyzeCoverageStopAreas
+from analyzers import AnalyzeCoverageStopArea
 import os
 
 pytestmark = pytest.mark.usefixtures("spark")
